requirements.txt
os
requests
tkinter
re
python-dotenv
gradio